Revert "T-iOS-426 - on smaller phones (for example iPhone 8) the Quick delete…"
This reverts commit a8583e226ca4683f86b086199f0559eb5f9f8a9f.
